Early Connection prototype
First version of the connection prototype being created.

I am thinking about how we want to manage connections and am hoping to
do so through connection objects that maintain connections between
computers. Servers will maintain a pool of these objects, and clients
will maintain a single connection. The data object will be used as a
layer around our change objects that will provide things like
timestamping etc. I think this will work. I wrote a custom json encoder
for the data objects (really just gets the __dict__ from the object so
it can be serialized. Timestamps are created as soon as the _Data is
created. Everything is nested in Connection right now as it seems that
it is unlikely that we should let these objects be used globably.