Rebuilt everything based on the new specs
Class/file dependancies are now correctly applied.

I went ahead and stubbed out the classes and their dependancies for now.
These will next go through prelim documentation. I want to get the docs
started now so we don't have to do them in the future. It is the next
step to be done. Since the heirarchy seems correct now we can also move
on to building the actual functionality. We basically now have the
supports in place to start moving forward.

Built first iteration of TransmissionData

The TransmissionData object is now built

This object contains two pieces of information currently. The first is
the time which the package was created, and the next is the data
maintained in the cash. This is allowed to be of arbitrary type - it can
literally be any python type. For our cases it will probably be a list
or a dictionary - either way it will work. The second is a time stamp.
It can be set during instantiation, but if not it will default to the
current time during which the package was created. This will allow for
sending packages easily as creation of a new package, and rebuilding a
transmitted package is done with the constructor and nothing else.