Built the custome json serializer
The TransmissionDataSerializer is now built and working.

By using the functions in this class it is possible to both Encode a
TransmissionData object via dumps(obj, cls=TransmissionDataSerializer).
To undo this loads(str, object_hook=TransmissionDataSerializer.Decode)
will give return the rebuilt object.

This means that we are now ready to start working with sending these
objects we can garuntee that the we will get back an identical object to
what we sent. These need unit tests - that is the next step.

Unit tests for TransmissionDataSerializer

I have added the first tests for the TransmissionDataSerializer. I am
playing catchup a little and need to get these up and some commenting. I
need to get my act together and write comments as I go and design tests
before I write code!!!