Unit testing for TransmissionDataSerializer Done
The unit tests are done and pass, updated various other things.

There are now working unit tests for the Serializer, this caught a few
errors which was nice. The current tests are for the following
Serializing TransmissionData with automatic Time stamping,
SerializingData with a predefined timestamp, Deserializing a
TransmissionData object.

Changed data members of TransmissionData to fit naming convention -
publict members are capitalized.

Biggest change to the serializer was that the Decode function now will
allow for decoding TransmissionData objects containing dictionaries and
lists. This means that nested dictionaries can't be under the key data
as it is sort of a reserved word.

I need to update the tester next.