Updates to be more pythonic, cleaned unittests
Better fits PEP 8 standards. Unit tests fit the stds from Pylons (much
cleaner and safer).

Shortened some calls in the TransmissionDataSerializerTests to use self.
This is still readable and lowers the line length to under chars.  Moved
necessary imports into the functions from which they are utilized which
is best practised so that if there is an import error it will cause a
failed test instead of halting the testing process.

Changed checking for None to use the is keyword instead of equality
checks/boolean checks. This was only in one place - good to catch this
early!

Changed TransmissionData to _TransmissionData

Updated the TransmissionData module to fit the updated UML for
Connections.

All tests still pass.

Added the custom loads and dumps

The custom loads and dumps have been implemented and tested.

These functions will act as an abstraction to the built in JSON module.
They are specialized for TransmissionData serializing and deserializing.
Tests are passing right now, this should be working ^_^

Updated the sphinx files

Moved sphinx files so that they don't clutter everything.

All of the new modules and tests are documented and sphinx is properly
building their pages.