Created Testing script
Testing script has been created to discover all the tests in our project
in the unit_test package. It is naive and just runs all tests and
displays the result. This currently is just run from the cmd line but
should allow for running any tests we create as long as they are
1) in the unit_test package
2) file names match the pattern *_tests.py