fixed travis-CI image
fixed the incorrect use of travis' image

The main repo's build status is at the top, and anyone can add their
status to to forks list.
this list also is cool because it links to the Travis-CI, shows the
build status of the fork, and also links directly to the fork.